# docs/swagger.yaml
openapi: 3.0.3
info:
  title: LLM-Lite Underwriter API
  description: |
    파라메트릭 보험 자동 인수심사 API 시스템
    
    ## 주요 기능
    - 자연어 입력을 통한 보험상품 설계
    - 비동기 작업 처리 및 실시간 진행률 추적
    - Webhook/파일/메시지큐를 통한 결과 전송
    - 배치 처리 지원
    
    ## 사용 흐름
    1. `/underwrite` 엔드포인트로 인수심사 요청 제출
    2. 반환된 `task_id`로 `/tasks/{task_id}` 엔드포인트에서 진행상황 확인
    3. 처리 완료 시 설정된 콜백 방식으로 결과 수신
    
  version: 1.0.0
  contact:
    name: LLM-Lite Development Team
    email: support@llm-lite.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000/api/v1
    description: 개발 서버
  - url: https://api.llm-lite.com/v1
    description: 프로덕션 서버
...
paths:
...
  /tasks/{task_id}:
    get:
      tags:
        - Tasks
      summary: 작업 상태 조회
      description: |
        task_id에 해당하는 작업의 상세 상태, 진행률, 결과 등을 반환합니다.
        
        **상태 종류:**
        - `accepted`: 요청 접수
        - `queued`: 큐 대기 중
        - `processing`: 처리 중
        - `completed`: 완료
        - `failed`: 실패
        - `cancelled`: 취소됨
        
      operationId: getTaskStatus
      parameters:
        - name: task_id
          in: path
          required: true
          description: 작업 고유 ID
          schema:
            type: string
            minLength: 10
            maxLength: 100
          example: "underwriter_20250121_143052_a1b2c3d4"
      responses:
        '200':
          description: 작업 상태 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskStatusResponse'
              examples:
                processing:
                  summary: 처리 중인 작업
                  value:
                    task_id: "underwriter_20250121_143052_a1b2c3d4"
                    request_id: "typhoon_001"
                    status: "processing"
                    priority: "high"
                    created_at: "2025-01-21T14:30:52.123456Z"
                    started_at: "2025-01-21T14:30:55.789012Z"
                    processing_duration: 45.2
                    progress:
                      percentage: 65.0
                      current_step: "pricing_calculation"
                      completed_steps: ["peril_canvas", "prior_extraction", "scenario_generation"]
                      total_steps: 5
                    estimated_completion: "2025-01-21T14:32:10.123456Z"
                completed:
                  summary: 완료된 작업
                  value:
                    task_id: "underwriter_20250121_143052_a1b2c3d4"
                    request_id: "typhoon_001"
                    status: "completed"
                    priority: "high"
                    created_at: "2025-01-21T14:30:52.123456Z"
                    started_at: "2025-01-21T14:30:55.789012Z"
                    completed_at: "2025-01-21T14:32:15.654321Z"
                    processing_duration: 83.5
                    progress:
                      percentage: 100.0
                      current_step: "completed"
                      completed_steps: ["peril_canvas", "prior_extraction", "scenario_generation", "pricing_calculation", "pricing_report"]
                      total_steps: 5
                    result:
                      status: "success"
                      expected_loss: 15000000
                      gross_premium: 18750000
                      risk_level: "medium"
                      loss_ratio: 0.8
                      coefficient_of_variation: 1.25
                      var_99: 85000000
                      tvar_99: 92000000
                      simulation_years: 1000
                      validation_passed: true
                      event_type: "태풍"
                      recommendation: "추천: 보험료 조정 후 인수 가능"
                    callback_status: "success"
        '404':
          description: 작업을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 서버 내부 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Tasks
      summary: 작업 취소
      description: |
        처리 중이거나 대기 중인 작업을 취소합니다.
        이미 완료되었거나 실패한 작업은 취소할 수 없습니다.
      operationId: cancelTask
      parameters:
        - name: task_id
          in: path
          required: true
          description: 작업 고유 ID
          schema:
            type: string
            minLength: 10
            maxLength: 100
      responses:
        '204':
          description: 작업 취소 성공
        '404':
          description: 작업을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 취소할 수 없는 상태 (이미 완료/실패)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /tasks:
    get:
      tags:
        - Tasks
      summary: 작업 목록 조회
      description: |
        시스템의 작업들을 조회합니다. 상태별 필터링과 페이징을 지원합니다.
      operationId: listTasks
      parameters:
        - name: status_filter
          in: query
          description: 상태별 필터
          schema:
            $ref: '#/components/schemas/TaskStatus'
        - name: limit
          in: query
          description: 반환할 작업 수 제한
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 50
        - name: include_completed
          in: query
          description: 완료된 작업 포함 여부
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: 작업 목록 조회 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  tasks:
                    type: array
                    items:
                      $ref: '#/components/schemas/TaskSummary'
                  total_count:
                    type: integer
                  filters:
                    type: object
                    properties:
                      status:
                        type: string
                        nullable: true
                      limit:
                        type: integer
                      include_completed:
                        type: boolean
                  timestamp:
                    type: string
                    format: date-time
  /tasks/{task_id}/retry:
    post:
      tags:
        - Tasks
      summary: 실패한 작업 재시도
      description: |
        실패한 작업을 다시 큐에 추가하여 재처리합니다.
        새로운 task_id가 생성됩니다.
      operationId: retryFailedTask
      parameters:
        - name: task_id
          in: path
          required: true
          description: 실패한 작업의 ID
          schema:
            type: string
            minLength: 10
            maxLength: 100
      responses:
        '202':
          description: 재시도 요청 접수 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Task retry submitted successfully"
                  original_task_id:
                    type: string
                    example: "underwriter_20250121_143052_a1b2c3d4"
                  new_task_id:
                    type: string
                    example: "underwriter_20250121_144052_b2c3d4e5"
                  status_url:
                    type: string
                    example: "/api/v1/tasks/underwriter_20250121_144052_b2c3d4e5"
                  timestamp:
                    type: string
                    format: date-time
        '404':
          description: 작업을 찾을 수 없음
        '409':
          description: 재시도할 수 없는 상태 (실패 상태가 아님)
...
components:
  schemas:
...
    TaskStatusResponse:
      type: object
      properties:
        task_id:
          type: string
          example: "underwriter_20250121_143052_a1b2c3d4"
        request_id:
          type: string
          example: "typhoon_001"
        status:
          $ref: '#/components/schemas/TaskStatus'
        priority:
          $ref: '#/components/schemas/TaskPriority'
        created_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true
        processing_duration:
          type: number
          format: float
          description: 처리 소요 시간 (초)
          nullable: true
          example: 83.5
        progress:
          $ref: '#/components/schemas/TaskProgress'
        estimated_completion:
          type: string
          format: date-time
          nullable: true
        result:
          $ref: '#/components/schemas/UnderwriteResult'
        error:
          type: string
          nullable: true
        callback_status:
          $ref: '#/components/schemas/CallbackStatus'
        callback_attempts:
          type: integer
          default: 0
        last_callback_attempt:
          type: string
          format: date-time
          nullable: true
    TaskSummary:
      type: object
      properties:
        task_id:
          type: string
        request_id:
          type: string
        status:
          type: string
        priority:
          type: string
        created_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true
        processing_duration:
          type: number
          format: float
          nullable: true
        progress_percentage:
          type: number
          format: float
        current_step:
          type: string
          nullable: true
        callback_status:
          type: string
        has_result:
          type: boolean
        has_error:
          type: boolean
    TaskProgress:
      type: object
      properties:
        percentage:
          type: number
          format: float
          minimum: 0.0
          maximum: 100.0
          description: 진행률 백분율
          example: 65.0
        current_step:
          type: string
          description: 현재 처리 단계
          nullable: true
          example: "pricing_calculation"
        completed_steps:
          type: array
          items:
            type: string
          description: 완료된 단계들
          example: ["peril_canvas", "prior_extraction", "scenario_generation"]
        total_steps:
          type: integer
          description: 전체 단계 수
          default: 5
          example: 5
        steps:
          type: array
          items:
            $ref: '#/components/schemas/ProcessingStep'
    ProcessingStep:
      type: object
      properties:
        name:
          type: string
          description: 단계명
          example: "peril_canvas"
        status:
          type: string
          description: 단계 상태
          enum: ["pending", "processing", "completed", "failed"]
          example: "completed"
        started_at:
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true
    UnderwriteResult:
      type: object
      properties:
        status:
          type: string
          description: 처리 결과 상태
          enum: ["success", "error"]
          example: "success"
        expected_loss:
          type: number
          description: 기대 손실액 (원)
          nullable: true
          example: 15000000
        gross_premium:
          type: number
          description: 총 보험료 (원)
          nullable: true
          example: 18750000
        risk_level:
          type: string
          description: 리스크 레벨
          enum: ["low", "medium", "high", "very_high"]
          nullable: true
          example: "medium"
        loss_ratio:
          type: number
          format: float
          description: 손해율
          nullable: true
          example: 0.8
        coefficient_of_variation:
          type: number
          format: float
          description: 변동계수
          nullable: true
          example: 1.25
        var_99:
          type: number
          description: 99% VaR (Value at Risk)
          nullable: true
          example: 85000000
        tvar_99:
          type: number
          description: 99% TVaR (Tail Value at Risk)
          nullable: true
          example: 92000000
        simulation_years:
          type: integer
          description: 시뮬레이션 연수
          example: 1000
        validation_passed:
          type: boolean
          description: 검증 통과 여부
          default: false
        validation_checks:
          type: array
          items:
            type: string
          description: 검증 항목들
          nullable: true
        alerts:
          type: array
          items:
            type: string
          description: 경고 메시지들
          nullable: true
        event_type:
          type: string
          description: 위험 사건 유형
          nullable: true
          example: "태풍"
        recommendation:
          type: string
          description: 인수심사 권고사항
          nullable: true
          example: "추천: 보험료 조정 후 인수 가능"
        executive_summary:
          type: string
          description: 경영진 요약
          nullable: true
        audit_trail:
          type: array
          items:
            type: object
          description: 감사 추적 정보
          nullable: true
        error:
          type: string
          description: 에러 메시지 (실패 시)
          nullable: true
...
    ErrorResponse:
      type: object
      properties:
        error_code:
          type: string
          description: 에러 코드
          example: "VALIDATION_ERROR"
        message:
          type: string
          description: 에러 메시지
          example: "Request validation failed"
        details:
          type: object
          description: 상세 에러 정보
          nullable: true
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
          nullable: true

    # 열거형 타입들
    TaskStatus:
      type: string
      enum:
        - accepted
        - queued
        - processing
        - completed
        - failed
        - cancelled
      description: |
        작업 상태:
        - `accepted`: 요청 접수
        - `queued`: 큐 대기 중
        - `processing`: 처리 중
        - `completed`: 완료
        - `failed`: 실패
        - `cancelled`: 취소됨
    TaskPriority:
      type: string
      enum:
        - low
        - normal
        - high
        - urgent
      default: normal
      description: |
        작업 우선순위:
        - `low`: 낮음
        - `normal`: 보통
        - `high`: 높음
        - `urgent`: 긴급
...
    CallbackStatus:
      type: string
      enum:
        - pending
        - in_progress
        - success
        - failed
      description: |
        콜백 상태:
        - `pending`: 대기 중
        - `in_progress`: 진행 중
        - `success`: 성공
        - `failed`: 실패
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API 키 인증 (선택사항, 환경에 따라 활성화)
security:
  - ApiKeyAuth: []
  - {}  # 인증 없이도 접근 가능